+++ AppVersion.cs
@@ -2,7 +2,7 @@ namespace Map3d
 {
     internal static class AppVersion
     {
-        public const string BepInSemVer = "0.3.3";
-        public const string DisplayVersion = "0.3.3";
+        public const string BepInSemVer = "0.3.4";
+        public const string DisplayVersion = "0.3.4";
     }
 }